% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Center a vector around a specified value}
\usage{
center(data, ..., center = 0)
}
\arguments{
\item{data}{The single-dimension, numeric data vector to center}

\item{center}{The numeric scalar around which the data shall be arranged, defaults to 0, named}
}
\value{
centered_data The centered numeric data set
}
\description{
Take a center value and rearrange data points around it regarding the mean.
}
\examples{
  center(1:3)
  # with default center should result in [1] -1  0  1

  center(c(1, 2, 3), 10)
  # should result in [1]  9 10 11
}
