% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Normalize a data set}
\usage{
rescale(v, ..., low = 0, high = 1)
}
\arguments{
\item{v}{The numeric, single-dimension vector which should be normalized}

\item{...}{any other paramter}

\item{low}{The scalar lower bound for normalization}

\item{high}{The scalar upper bound for normalization}
}
\value{
The normalized version of `v`
}
\description{
Normalize a data set
}
\examples{
rescale(c(0, 1, 2, 3, 4), low = 2, high = 10)
# should result in [1]  2  4  6  8 10

rescale(c(0, 1, 2, 3, 4))
# should result in [1] 0.00 0.25 0.50 0.75 1.00
}
